Change acquire search order to be sequential
DXVK has a bug where if the same swapchain image index is acquired consecutively, it'll break synchronization on their side. While this is a bug with DXVK, it's a general assumption that works fine in other presentation frameworks and other software may indirectly rely on this behavior as well so this was fixed in our compositor instead.