Delete CassiaSrv and existing JNI interface
As we're moving to a fundamentally different paradigm for how Cassia C++ code is structured with the decision to move from our own Wine DRV to using a stack dependent on external libraries/programs (X11, gamescope, etc) with any native code relevant to them in `cassiaext` (including the server bits). All the existing code can be purged and replaced with newer glue code to launch everything from `cassiaext` as well as managing the runtime.