Implement `WineContext` in C++
A class for consolidating all state around Wine for a certain prefix, it's responsible for launching `wineserver` and initializing the prefix with `wineboot.exe` then handling the lifetime of the whole session by terminating everything in its destructor. Additionally, a convenience function for launching any wine processes is included to avoid needing to specify the runtime and prefix path for every invocation.