Change `Process` lifetime semantics for safety
The implicit killing of any child process in the `Process` destructor was deemed to be unsafe, yet the lifetime of a child process not being attached to the RAII `Process` object without explicitly doing so was also fallible. As a result, a similar approach to `std::thread` has been adopted with the child process needing to be detached or stopped by the time the destructor is called or it will result in `std::terminate` being called.