Implement `InputHandler` for surface input events
All motion/key events we receive from the surface need to be translated into an input events that can be consumed by the compositor. The `InputHandler` class is designed for this, it currently takes in any event from the surface and translates it into the appropriate format for a compositor. In lieu of a compositor, it currently only logs the input while not actually passing it through.