Add Gradle task to build and package `cassiaext`
We needed a system to package an output tarball produced by `cassiaext`, this needed to function with WSL and support using the Nix derivation included with `cassiext` so a Gradle task was used to call all necessary commands. After invoking a build with `ninja`, it packages the output tarball as an Android asset named `cassiaext.tar.gz`.